First try at generating a CNF for Costas arrays
The current script has all the elements and P6 plays nicely, but I must
have made an error somewhere in the deduction of the axioms, probably,
because the generated CNF for 3-Costas arrays is UNSAT, but there are
four 3-Costas arrays.

I got interested in this topic this afternoon when my colleague next
door, Ali Ardalani, gave a talk about it.