Allow large CNFs to be written
The DIMACS cnf format requires the number of variables and clauses in
the header. It is difficult to know the number of clauses in advance
for this problem, so previously we collected all of the axioms into an
array before writing the header. This requires too much memory when
the requested order of Costas arrays rises (the current formulation
uses about N**6 clauses).

Now a padding of 30 decimal '0's is written into the header at first
and the axioms are lazily generated and counted. At the end, the exact
number of axioms is patched into the header. The sharpSAT solver
which I tested accepts leading zeros. This change requires giving
the script a destination filename though, so output to stdout is not
supported anymore.

Alternatives would include writing just the axioms without the header
and prepending the header afterwards, either (1) with a separate script
or (2) on-the-fly when feeding it to the SAT solver. I don't like
either of these alternatives because (1) would need to create a
temporary copy of the file for prepending something, and (2) would
not lead to a self-contained DIMACS file and requires counting the
lines every time the file is used, and the file can be multiple GiB
in size.

Also add a TODO file with thoughts about an alternate formulation of
the Costas array definition, and another one specialized for the
interesting case of N=32.